refactor(typescript): group server deduplication modules
